user service: use grpc

+++ services/user/model.go
@@ -5,7 +5,7 @@ import (
 )
 
 type Auth struct {
-	UserID    uint   `gorm:"AUTO_INCREMENT;NOT NULL;PRIMARY_KEY"`
+	UserID    uint64 `gorm:"AUTO_INCREMENT;NOT NULL;PRIMARY_KEY"`
 	Username  string `gorm:"UNIQUE_INDEX;type:varchar(128)"`
 	Password  string
 	CreatedAt time.Time
@@ -13,7 +13,7 @@ type Auth struct {
 }
 
 type Profile struct {
-	UserID   uint   `gorm:"AUTO_INCREMENT:false;NOT NULL;PRIMARY_KEY"`
+	UserID   uint64 `gorm:"AUTO_INCREMENT:false;NOT NULL;PRIMARY_KEY"`
 	Username string `gorm:"UNIQUE_INDEX;type:varchar(128)"`
 	Bio      string
 	AvatarID string
